Trivial refactoring
